+++ src/components/date-picker/date-picker.tsx
@@ -24,7 +24,13 @@ import { HeadingLevel } from "../functional/Heading";
 
 import { DateRangeChange } from "./interfaces";
 import { HEADING_LEVEL, TEXT } from "./resources";
-import { connectLocalized, disconnectLocalized, LocalizedComponent } from "../../utils/locale";
+import {
+  connectLocalized,
+  disconnectLocalized,
+  LocalizedComponent,
+  NumberingSystem,
+  numberStringFormatter
+} from "../../utils/locale";
 
 @Component({
   assetsDirs: ["assets"],
@@ -155,6 +161,12 @@ export class DatePicker implements LocalizedComponent {
    */
   @Prop() locale?: string;
 
+  /**
+   * Specifies the Unicode numeral system used by the component for localization. This property cannot be dynamically changed.
+   *
+   */
+  @Prop({ reflect: true }) numberingSystem?: NumberingSystem;
+
   /** specify the scale of the date picker */
   @Prop({ reflect: true }) scale: "s" | "m" | "l" = "m";
 
@@ -238,6 +250,12 @@ export class DatePicker implements LocalizedComponent {
     if (this.max) {
       this.maxAsDate = dateFromISO(this.max);
     }
+
+    numberStringFormatter.numberFormatOptions = {
+      numberingSystem: this.numberingSystem,
+      locale: this.effectiveLocale,
+      useGrouping: false
+    };
   }
 
   disconnectedCallback(): void {
@@ -347,6 +365,12 @@ export class DatePicker implements LocalizedComponent {
     if (!Build.isBrowser) {
       return;
     }
+
+    numberStringFormatter.numberFormatOptions = {
+      numberingSystem: this.numberingSystem,
+      locale: this.effectiveLocale,
+      useGrouping: false
+    };
 
     this.localeData = await getLocaleData(this.effectiveLocale);
   }