+++ src/components/input-date-picker/input-date-picker.tsx
@@ -56,8 +56,15 @@ import {
   connectOpenCloseComponent,
   disconnectOpenCloseComponent
 } from "../../utils/openCloseComponent";
-import { connectLocalized, disconnectLocalized, LocalizedComponent } from "../../utils/locale";
+import {
+  connectLocalized,
+  disconnectLocalized,
+  LocalizedComponent,
+  NumberingSystem,
+  numberStringFormatter
+} from "../../utils/locale";
 import { debounce } from "lodash-es";
+import { numberKeys } from "../../utils/key";
 
 @Component({
   tag: "calcite-input-date-picker",
@@ -86,12 +93,12 @@ export class InputDatePicker
   //
   //--------------------------------------------------------------------------
   /**
-   * When true, interaction is prevented and the component is displayed with lower opacity.
+   * When `true`, interaction is prevented and the component is displayed with lower opacity.
    */
   @Prop({ reflect: true }) disabled = false;
 
   /**
-   * When true, the component's value can be read, but controls are not accessible and the value cannot be modified.
+   * When `true`, the component's value can be read, but controls are not accessible and the value cannot be modified.
    *
    * @mdn [readOnly](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/readonly)
    */
@@ -190,7 +197,7 @@ export class InputDatePicker
   }
 
   /**
-   * When true, the component is active.
+   * When `true`, the component is active.
    *
    * @deprecated use `open` instead.
    */
@@ -201,7 +208,7 @@ export class InputDatePicker
     this.open = value;
   }
 
-  /** When true, displays the `calcite-date-picker` component. */
+  /** When `true`, displays the `calcite-date-picker` component. */
   @Prop({ mutable: true, reflect: true }) open = false;
 
   @Watch("open")
@@ -250,6 +257,12 @@ export class InputDatePicker
    */
   @Prop() locale?: string;
 
+  /**
+   * Specifies the Unicode numeral system used by the component for localization. This property cannot be dynamically changed.
+   *
+   */
+  @Prop({ reflect: true }) numberingSystem?: NumberingSystem;
+
   /** Specifies the size of the component. */
   @Prop({ reflect: true }) scale: "s" | "m" | "l" = "m";
 
@@ -260,11 +273,11 @@ export class InputDatePicker
    */
   @Prop({ reflect: true }) placement: MenuPlacement = defaultMenuPlacement;
 
-  /** When true, activates a range for the component. */
+  /** When `true`, activates a range for the component. */
   @Prop({ reflect: true }) range = false;
 
   /**
-   * When true, the component must have a value in order for the form to submit.
+   * When `true`, the component must have a value in order for the form to submit.
    *
    * @internal
    */
@@ -300,7 +313,7 @@ export class InputDatePicker
   }
 
   /**
-   * When true, disables the default behavior on the third click of narrowing or extending the range.
+   * When `true`, disables the default behavior on the third click of narrowing or extending the range.
    * Instead starts a new range.
    */
   @Prop() proximitySelectionDisabled = false;
@@ -456,6 +469,12 @@ export class InputDatePicker
 
     this.setFilteredPlacements();
     this.debouncedReposition();
+
+    numberStringFormatter.numberFormatOptions = {
+      numberingSystem: this.numberingSystem,
+      locale: this.effectiveLocale,
+      useGrouping: false
+    };
   }
 
   async componentWillLoad(): Promise<void> {
@@ -496,7 +515,9 @@ export class InputDatePicker
                   disabled={disabled}
                   icon="calendar"
                   label={getLabelText(this)}
+                  lang={this.effectiveLocale}
                   number-button-type="none"
+                  numberingSystem={this.numberingSystem}
                   onCalciteInputInput={this.calciteInternalInputInputHandler}
                   onCalciteInternalInputBlur={this.calciteInternalInputBlurHandler}
                   onCalciteInternalInputFocus={this.startInputFocus}
@@ -533,10 +554,12 @@ export class InputDatePicker
                   intlNextMonth={this.intlNextMonth}
                   intlPrevMonth={this.intlPrevMonth}
                   intlYear={this.intlYear}
+                  lang={this.effectiveLocale}
                   max={this.max}
                   maxAsDate={this.maxAsDate}
                   min={this.min}
                   minAsDate={this.minAsDate}
+                  numberingSystem={this.numberingSystem}
                   onCalciteDatePickerChange={this.handleDateChange}
                   onCalciteDatePickerRangeChange={this.handleDateRangeChange}
                   proximitySelectionDisabled={this.proximitySelectionDisabled}
@@ -568,7 +591,9 @@ export class InputDatePicker
                   }}
                   disabled={disabled}
                   icon="calendar"
+                  lang={this.effectiveLocale}
                   number-button-type="none"
+                  numberingSystem={this.numberingSystem}
                   onCalciteInputInput={this.calciteInternalInputInputHandler}
                   onCalciteInternalInputBlur={this.calciteInternalInputBlurHandler}
                   onCalciteInternalInputFocus={this.endInputFocus}
@@ -964,4 +989,31 @@ export class InputDatePicker
       `The specified value "${value}" does not conform to the required format, "yyyy-MM-dd".`
     );
   }
+
+  private commonDateSeparators = [".", "-", "/"];
+
+  private formatNumerals = (value: string): string =>
+    value
+      ? value
+          .split("")
+          .map((char: string) =>
+            // convert common separators to the locale's
+            this.commonDateSeparators.includes(char)
+              ? this.localeData.separator
+              : numberKeys.includes(char)
+              ? numberStringFormatter.numberFormatter.format(Number(char))
+              : char
+          )
+          .join("")
+      : "";
+
+  private parseNumerals = (value: string): string =>
+    value
+      ? value
+          .split("")
+          .map((char: string) =>
+            numberKeys.includes(char) ? numberStringFormatter.delocalize(char) : char
+          )
+          .join("")
+      : "";
 }